Refactor platform and device fixtures to a seperate file
