+++ test/conformance/platform/urPlatformGet.cpp
@@ -3,9 +3,9 @@
 // See LICENSE.TXT
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "fixtures.h"
+#include "uur/fixtures.h"
 
-using urPlatformGetTest = uur::platform::urTest;
+using urPlatformGetTest = uur::urTest;
 
 TEST_F(urPlatformGetTest, Success) {
     uint32_t count;
