+++ test/conformance/platform/urPlatformGetApiVersion.cpp
@@ -3,9 +3,9 @@
 // See LICENSE.TXT
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "fixtures.h"
+#include "uur/fixtures.h"
 
-using urPlatformGetApiVersionTest = uur::platform::urPlatformTest;
+using urPlatformGetApiVersionTest = uur::urPlatformTest;
 
 TEST_F(urPlatformGetApiVersionTest, Success) {
     ur_api_version_t version;
